Some updates for startup

+++ install/setup_links.sh
@@ -1,11 +1,21 @@
 #! /bin/bash
 
-set -e
+set -x
 
-mv ~/.config ~/_temp_config
-ln -s ~/git/config_manager/xdg_config ~/.config
-cp -r ~/_temp_config/* ~/.config/
-rm -rf ~/_temp_config
+if [ -d "~/.config" ]; then
+    MOVE_CONFIG=true
+fi
+
+if [ $MOVE_CONFIG ]; then
+    mv ~/.config ~/_temp_config
+fi
+
+ln -sv ~/git/config_manager/xdg_config ~/.config
+
+if [ $MOVE_CONFIG ]; then
+    cp -r ~/_temp_config/* ~/.config/
+    rm -rf ~/_temp_config
+fi
 
 # Home files
-ln -s ~/git/config_manager/home/.zshenv ~/.zshenv
+ln -vs ~/git/config_manager/home/.zshenv ~/.zshenv
